+++ lib/ask/web_fetch.rb
@@ -3,5 +3,6 @@
 require_relative 'web_fetch/version'
 require_relative 'web_fetch/backend'
 require_relative 'web_fetch/backends/local'
+require_relative 'web_fetch/backends/crawl4ai'
 require_relative 'web_fetch/backends/jina'
 require_relative 'web_fetch/tool'